Changes format string from %lld to %PRIu64 in several tests.
6 liburing tests failed to build on powerpc due to a 'long' versus
'long long' mismatch. Changing the identifer from %lld to %PRIu64
and casting the value to uint64_t fixes this.

+++ test/cq-peek-batch.c
@@ -74,7 +74,8 @@ int main(int argc, char *argv[])
 	CHECK_BATCH(&ring, got, cqes, 4, 4);
 	for (i=0;i<4;i++) {
 		if (i != cqes[i]->user_data) {
-			printf("Got user_data %lld, expected %d\n", cqes[i]->user_data, i);
+			printf("Got user_data %" PRIu64 ", expected %d\n",
+				(uint64_t) cqes[i]->user_data, i);
 			goto err;
 		}
 	}
@@ -86,7 +87,8 @@ int main(int argc, char *argv[])
 	CHECK_BATCH(&ring, got, cqes, 4, 4);
 	for (i=0;i<4;i++) {
 		if (i + 4 != cqes[i]->user_data) {
-			printf("Got user_data %lld, expected %d\n", cqes[i]->user_data, i + 4);
+			printf("Got user_data %" PRIu64 ", expected %d\n",
+				(uint64_t) cqes[i]->user_data, i + 4);
 			goto err;
 		}
 	}
